agents/calculator/pyproject.toml
[project]
name = "ust-calculator-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "a2a-sdk[http-server]>=0.3.7",
  "uvicorn>=0.30",
  "mcp>=1.15.0",
  "httpx>=0.27"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"